+++ Modules-Global/Buzzer/alarm-rpi.py
@@ -1,6 +1,6 @@
 #Author: Jesimar da Silva Arantes
 #Date: 15/03/2018
-#Last Update: 06/11/2018
+#Last Update: 07/11/2018
 #Description: Code that turns on the alarm when the buzzer is turned on several times on the Raspberry Pi.
 #Descricao: Codigo que liga soa um alarme ao ligar o buzzer diversas vezes na Raspberry Pi.
 
@@ -10,7 +10,9 @@ import sys
 
 GPIO.setmode(GPIO.BOARD)
 
-pin = int(sys.argv[1]) # sends the signal (pin 40)
+GPIO.setwarnings(False)
+
+pin = int(sys.argv[1]) # sends the signal (pin 38 - called too BCM20)
 
 GPIO.setup(pin, GPIO.OUT)