+++ Modules-Global/LED/blink-led-rpi.py
@@ -1,6 +1,6 @@
 #Author: Jesimar da Silva Arantes
 #Date: 29/03/2018
-#Last Update: 06/11/2018
+#Last Update: 07/11/2018
 #Description: Code that blink the led. 
 #Descricao: Codigo que pisca o led.
 
@@ -10,7 +10,9 @@ import sys
 
 GPIO.setmode(GPIO.BOARD)
 
-pin = int(sys.argv[1]) # sends the signal (pin 40)
+GPIO.setwarnings(False)
+
+pin = int(sys.argv[1]) # sends the signal (pin 36 - called too BCM16)
 delay = float(sys.argv[2]) # delay time between blink
 
 GPIO.setup(pin, GPIO.OUT)