+++ Modules-Global/Parachute/open-parachute-rpi.py
@@ -1,6 +1,6 @@
 #Author: Jesimar da Silva Arantes
 #Date: 29/03/2018
-#Last Update: 06/11/2018
+#Last Update: 07/11/2018
 #Description: Code that open the parachute. 
 #Descricao: Codigo que abre o paraquedas. 
 
@@ -9,7 +9,9 @@ import sys
 
 GPIO.setmode(GPIO.BOARD)
 
-pin = int(sys.argv[1]) # sends the signal (pin 40)
+GPIO.setwarnings(False)
+
+pin = int(sys.argv[1]) # sends the signal (pin 37 - called too BCM26)
 
 GPIO.setup(pin, GPIO.OUT)