+++ Modules-Global/Spraying/close-spraying-rpi.py
@@ -1,6 +1,6 @@
 #Author: Jesimar da Silva Arantes
 #Date: 02/04/2018
-#Last Update: 06/11/2018
+#Last Update: 07/11/2018
 #Description: Code that spraying. 
 #Descricao: Codigo que encerra a pulveriacao. 
 
@@ -9,7 +9,9 @@ import sys
 
 GPIO.setmode(GPIO.BOARD)
 
-pin = int(sys.argv[1]) # sends the signal (pin 40)
+GPIO.setwarnings(False)
+
+pin = int(sys.argv[1]) # sends the signal (pin 40 - called too BCM21)
 
 GPIO.setup(pin, GPIO.OUT)